+++ src/BDSOutputROOTEventHeader.cc
@@ -27,6 +27,7 @@ along with BDSIM.  If not, see <http://www.gnu.org/licenses/>.
 
 #include <ctime>
 #include <string>
+#include <vector>
 
 ClassImp(BDSOutputROOTEventHeader)
 
@@ -48,6 +49,8 @@ void BDSOutputROOTEventHeader::Flush()
   fileType      = "BDSIM";
   dataVersion   = 5; // update analysis/DataLoader.cc default when this changes
   // also in comparator/Compare.cc at top - EXPECTEDDATAVERSION
+  analysedFiles.clear();
+  combinedFiles.clear();
   
 #ifndef __ROOTDOUBLE__
   doublePrecisionOutput = false;
@@ -56,10 +59,13 @@ void BDSOutputROOTEventHeader::Flush()
 #endif
 }
 
-void BDSOutputROOTEventHeader::Fill()
+void BDSOutputROOTEventHeader::Fill(const std::vector<std::string>& analysedFilesIn,
+				    const std::vector<std::string>& combinedFilesIn)
 {
   time_t rawtime;
   time(&rawtime);
   timeStamp = std::string(ctime(&rawtime));
+  analysedFiles = analysedFilesIn;
+  combinedFiles = combinedFilesIn;
 }